Major Progress!
Completed the output and of the Neural Network(NN)! Currently the NN can achieve around 60-70% accuracy, which is higher than random!
I also improved the speed of InitialStart, fetching data and creating data structure for images and labels.
Next step is to make a GUI for the user to draw and image and return an output. 
In addition,  maybe create a visual for this neural network development.